Jetpack Connect: Only auto-auth connections started by *button-click*
Previously it was possible to bypass the TOS and 'Approve' button
in the jetpack connect process by arriving directly at a URL
such as wordpress.com/jetpack/connect?url=somedotorgsite.

Change the auto-approve logic so that only a button-click in
Calypso (which acknowledges TOS) will cause the 'Approve' button
to be skipped at the Authorize step.

Necessary Changes:
* Move the persistSession call from the action to the onClick handlers
* Move cleanUrl() to utils to make it available to jetpack-new-site component
* Remove special handling for mobile apps now that arriving with ?url query does not auto-approve
